I noticed that double is overkill for pixel conversion operations.
Generic Color now uses floats instead of doubles, because doubles
are overkill for the color conversions.

If 16 bit per channels like R16G16B16A16 formats are used I might
need to use doubles for the generic color.

+++ src/Utilities/PixelFormat.h
@@ -8,7 +8,7 @@
 
 namespace Utilities {
 
-    typedef double channel_fp;
+    typedef float channel_fp;
     typedef uint16_t palette_index;
 
     class PixelFormat {
